+++ example/src/main/kotlin/com/animals/AbstractAnimal.kt
@@ -1,0 +1,7 @@
+package com.animals
+
+abstract class AbstractAnimal {
+    abstract val name: String
+    val isExtinct: Boolean = false
+    open fun makeSound(): String = ""
+}