+++ example/src/main/kotlin/com/animals/Cat.kt
@@ -1,0 +1,6 @@
+package com.animals
+
+class Cat: AbstractAnimal() {
+    override val name : String = "CAT"
+    override fun makeSound():String  =""
+}